w24table improvements

+++ CycleFinder.Models/Ephemeris/PlanetaryLine.cs
@@ -6,9 +6,9 @@ namespace CycleFinder.Models.Ephemeris
     public class PlanetaryLine
     {
         public Planet Planet { get; }
-        public IEnumerable<(DateTime Time, double Value)> Values { get; }
+        public IEnumerable<(DateTime Time, int Value)> Values { get; }
 
-        public PlanetaryLine(Planet planet, IEnumerable<(DateTime Time, double Value)> values)
+        public PlanetaryLine(Planet planet, IEnumerable<(DateTime Time, int Value)> values)
         {
             Planet = planet;
             Values = values;
